Extract filterLessonsBySearchTerm
